pyproject.toml
[project]
name = "shelfai"
version = "0.1.0"
description = "AI-powered grocery shelf analysis using computer vision and large language models for retail inventory management"
readme = "README.md"
authors = [
    { name = "Patrick Almeida", email = "patrickbalmeida@gmail.com" },
    { name = "Pedro Pereira", email = "devpereirapedro@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.53.0",
    "google-generativeai>=0.8.5",
    "numpy>=2.2.6",
    "openai>=1.86.0",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.1.0",
    "torch>=2.7.1",
    "ultralytics>=8.3.153",
]

[project.scripts]
shelfai = "ShelfAI.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "shelfai",
    "pytest>=8.3.5",
]

[tool.uv.sources]
shelfai = { workspace = true }
